chore(nix/api)!: remove _.[components].name
The superior implementation is the one that uses code generation to
create a `_/name/default.nix`, which is imported for `name = ...`.

This approach requires less manual work on the user's part, and no
duplication of information (the name is already given by the path of
`main.nix`).

Ideally, when the path of the derivation changes, the `generate` script should
handle the changes for the user seamlessly.